+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add subscription button to donation widget.
 - Added SCSS variable for background image for easier customization.
+- Allow to cancel payment.
 
 ### Changed